Move PIDFController 2 class into separate file and format
