+++ src/main/kotlin/org/sert2521/infiniterecharge2020/Main.kt
@@ -2,11 +2,9 @@ package org.sert2521.infiniterecharge2020
 
 import org.sert2521.infiniterecharge2020.drivetrain.Drivetrain
 import org.sert2521.infiniterecharge2020.powerhouse.PowerHouse
-import org.sert2521.sertain.events.whileAuto
 import org.sert2521.sertain.events.whileTeleop
 import org.sert2521.sertain.robot
 import org.sert2521.sertain.subsystems.add
-
 
 suspend fun main() = robot {
     println("Robot program starting")
