+++ src/main/kotlin/org/sert2521/infiniterecharge2020/drivetrain/Drivetrain.kt
@@ -3,6 +3,7 @@ package org.sert2521.infiniterecharge2020.drivetrain
 import com.ctre.phoenix.motorcontrol.ControlMode
 import com.kauailabs.navx.frc.AHRS
 import edu.wpi.first.wpilibj.I2C
+import kotlin.math.IEEErem
 import org.sert2521.infiniterecharge2020.MotorControllers
 import org.sert2521.sertain.coroutines.RobotScope
 import org.sert2521.sertain.motors.MotorController
@@ -11,7 +12,6 @@ import org.sert2521.sertain.telemetry.Table
 import org.sert2521.sertain.telemetry.TableEntry
 import org.sert2521.sertain.telemetry.linkTableEntry
 import org.sert2521.sertain.telemetry.withTableEntry
-import kotlin.math.IEEErem
 
 class Drivetrain : Subsystem("Drivetrain", ::controlDrivetrain) {
     val table = Table(name)
